agno-near-zcash/requirements.txt
# Core requirements
agno>=0.5.0
openai>=1.0.0
python-dotenv>=1.0.0
requests>=2.31.0
near-api-py>=0.5.0  # For NEAR blockchain interaction
base58>=2.1.1       # For key encoding
borsh>=0.3.1        # For data serialization

# Agent-specific requirements
duckduckgo-search>=3.0.0  # For market research capabilities
yfinance>=0.2.36          # For financial data

# Development dependencies
pytest>=7.0.0
black>=23.0.0
isort>=5.12.0

# Optional requirements
streamlit>=1.31.0  # For building a simple UI (optional) 